Improved: getProductIdentificationValue function shifted to utils/index.ts from utils/product/index.ts and made import changes wherever required
