+++ src/utils/index.ts
@@ -1,6 +1,5 @@
 import { toastController } from '@ionic/vue';
 import { DateTime } from "luxon";
-import productHelpers from './product'
 
 // TODO Use separate files for specific utilities
 
@@ -35,4 +34,25 @@ const handleDateTimeInput = (dateTimeValue: any) => {
   return DateTime.fromISO(dateTime).toMillis()
 }
 
-export { showToast, hasError, handleDateTimeInput, productHelpers }
+const getProductIdentificationValue = (productIdentifier: string, product: any) => {
+
+  // handled this case as on page load initially the data is not available, so not to execute furthur code
+  // untill product are not available
+  if(!Object.keys(product).length) {
+    return;
+  }
+
+  let value = product[productIdentifier]
+
+  // considered that the goodIdentification will always have values in the format "productIdentifier/value" and there will be no entry like "productIdentifier/"
+  const identification = product['goodIdentifications'].find((identification: string) => identification.startsWith(productIdentifier + "/"))
+
+  if(identification) {
+    const goodIdentification = identification.split('/')
+    value = goodIdentification[1]
+  }
+
+  return value;
+}
+
+export { showToast, hasError, handleDateTimeInput, getProductIdentificationValue }
